Add tests for NN yaml export
