--- petab_sciml/tmp_examples/data/nn_model1.yaml
@@ -1,164 +1,0 @@
-nn_model_id: model0
-inputs:
-- input_id: input0
-layers:
-- layer_id: conv1
-  layer_type: Conv2d
-  args:
-    stride:
-    - 1
-    - 1
-    padding:
-    - 0
-    - 0
-    dilation:
-    - 1
-    - 1
-    groups: 1
-    padding_mode: zeros
-    in_channels: 1
-    out_channels: 6
-    kernel_size:
-    - 5
-    - 5
-    bias: true
-- layer_id: conv2
-  layer_type: Conv2d
-  args:
-    stride:
-    - 1
-    - 1
-    padding:
-    - 0
-    - 0
-    dilation:
-    - 1
-    - 1
-    groups: 1
-    padding_mode: zeros
-    in_channels: 6
-    out_channels: 16
-    kernel_size:
-    - 5
-    - 5
-    bias: true
-- layer_id: fc1
-  layer_type: Linear
-  args:
-    in_features: 400
-    out_features: 120
-    bias: true
-- layer_id: fc2
-  layer_type: Linear
-  args:
-    in_features: 120
-    out_features: 84
-    bias: true
-- layer_id: fc3
-  layer_type: Linear
-  args:
-    in_features: 84
-    out_features: 10
-    bias: true
-forward:
-- name: input_1
-  op: placeholder
-  target: input
-  args: []
-  kwargs: {}
-- name: conv1
-  op: call_module
-  target: conv1
-  args:
-  - input_1
-  kwargs: {}
-- name: relu
-  op: call_function
-  target: relu
-  args:
-  - conv1
-  kwargs:
-    inplace: false
-- name: max_pool2d
-  op: call_function
-  target: max_pool2d
-  args:
-  - relu
-  - - 2
-    - 2
-  kwargs:
-    stride: null
-    padding: 0
-    dilation: 1
-    ceil_mode: false
-    return_indices: false
-- name: conv2
-  op: call_module
-  target: conv2
-  args:
-  - max_pool2d
-  kwargs: {}
-- name: relu_1
-  op: call_function
-  target: relu
-  args:
-  - conv2
-  kwargs:
-    inplace: false
-- name: max_pool2d_1
-  op: call_function
-  target: max_pool2d
-  args:
-  - relu_1
-  - 2
-  kwargs:
-    stride: null
-    padding: 0
-    dilation: 1
-    ceil_mode: false
-    return_indices: false
-- name: flatten
-  op: call_function
-  target: flatten
-  args:
-  - max_pool2d_1
-  - 1
-  kwargs: {}
-- name: fc1
-  op: call_module
-  target: fc1
-  args:
-  - flatten
-  kwargs: {}
-- name: relu_2
-  op: call_function
-  target: relu
-  args:
-  - fc1
-  kwargs:
-    inplace: false
-- name: fc2
-  op: call_module
-  target: fc2
-  args:
-  - relu_2
-  kwargs: {}
-- name: relu_3
-  op: call_function
-  target: relu
-  args:
-  - fc2
-  kwargs:
-    inplace: false
-- name: fc3
-  op: call_module
-  target: fc3
-  args:
-  - relu_3
-  kwargs: {}
-- name: output
-  op: output
-  target: output
-  args:
-  - fc3
-  kwargs: {}
