--- petab_sciml/tmp_examples/data2/nn_model1.yaml
@@ -1,49 +1,0 @@
-nn_model_id: model0
-inputs:
-- input_id: input0
-layers:
-- layer_id: conv1
-  layer_type: Conv2d
-  args:
-    stride:
-    - 1
-    - 1
-    padding:
-    - 0
-    - 0
-    dilation:
-    - 1
-    - 1
-    groups: 1
-    padding_mode: zeros
-    in_channels: 1
-    out_channels: 6
-    kernel_size:
-    - 5
-    - 5
-    bias: true
-forward:
-- name: input_1
-  op: placeholder
-  target: input
-  args: []
-  kwargs: {}
-- name: conv1
-  op: call_module
-  target: conv1
-  args:
-  - input_1
-  kwargs: {}
-- name: relu
-  op: call_function
-  target: relu
-  args:
-  - conv1
-  kwargs:
-    inplace: false
-- name: output
-  op: output
-  target: output
-  args:
-  - relu
-  kwargs: {}
