--- petab_sciml/tmp_examples/data3/nn_model0.yaml
@@ -1,80 +1,0 @@
-nn_model_id: model0
-inputs:
-- input_id: input0
-layers:
-- layer_id: layer1
-  layer_type: BatchNorm3d
-  args:
-    track_running_stats: true
-    momentum: 0.1
-    eps: 1.0e-05
-    num_features: 5
-    affine: true
-- layer_id: layer2
-  layer_type: InstanceNorm2d
-  args:
-    track_running_stats: false
-    momentum: 0.1
-    eps: 1.0e-05
-    num_features: 25
-    affine: false
-- layer_id: layer3
-  layer_type: BatchNorm1d
-  args:
-    track_running_stats: true
-    momentum: 0.1
-    eps: 1.0e-05
-    num_features: 125
-    affine: true
-- layer_id: flatten1
-  layer_type: Flatten
-  args:
-    start_dim: 1
-    end_dim: 2
-- layer_id: flatten2
-  layer_type: Flatten
-  args:
-    start_dim: 1
-    end_dim: 2
-forward:
-- name: input_1
-  op: placeholder
-  target: input
-  args: []
-  kwargs: {}
-- name: layer1
-  op: call_module
-  target: layer1
-  args:
-  - input_1
-  kwargs: {}
-- name: flatten1
-  op: call_module
-  target: flatten1
-  args:
-  - layer1
-  kwargs: {}
-- name: layer2
-  op: call_module
-  target: layer2
-  args:
-  - flatten1
-  kwargs: {}
-- name: flatten2
-  op: call_module
-  target: flatten2
-  args:
-  - layer2
-  kwargs: {}
-- name: layer3
-  op: call_module
-  target: layer3
-  args:
-  - flatten2
-  kwargs: {}
-- name: output
-  op: output
-  target: output
-  args:
-  - layer3
-  kwargs: {}
