--- petab_sciml/tmp_examples/data4/nn_model0.yaml
@@ -1,33 +1,0 @@
-nn_model_id: model0
-inputs:
-- input_id: input0
-layers:
-- layer_id: layer1
-  layer_type: RNN
-  args:
-    input_size: 5
-    hidden_size: 10
-    proj_size: 0
-    num_layers: 3
-    bias: true
-    batch_first: false
-    dropout: 0.0
-    bidirectional: false
-forward:
-- name: input_1
-  op: placeholder
-  target: input
-  args: []
-  kwargs: {}
-- name: layer1
-  op: call_module
-  target: layer1
-  args:
-  - input_1
-  kwargs: {}
-- name: output
-  op: output
-  target: output
-  args:
-  - layer1
-  kwargs: {}
