--- petab_sciml/tmp_examples/data5/nn_model0.yaml
@@ -1,72 +1,0 @@
-nn_model_id: model0
-inputs:
-- input_id: input0
-layers:
-- layer_id: norm1
-  layer_type: LayerNorm
-  args:
-    normalized_shape:
-    - 4
-    - 10
-    - 11
-    - 12
-    eps: 1.0e-05
-    elementwise_affine: true
-- layer_id: layer1
-  layer_type: Conv3d
-  args:
-    stride:
-    - 1
-    - 1
-    - 1
-    padding:
-    - 0
-    - 0
-    - 0
-    dilation:
-    - 1
-    - 1
-    - 1
-    groups: 1
-    padding_mode: zeros
-    in_channels: 4
-    out_channels: 1
-    kernel_size:
-    - 5
-    - 5
-    - 5
-- layer_id: flatten1
-  layer_type: Flatten
-  args:
-    start_dim: 1
-    end_dim: -1
-forward:
-- name: input_1
-  op: placeholder
-  target: input
-  args: []
-  kwargs: {}
-- name: norm1
-  op: call_module
-  target: norm1
-  args:
-  - input_1
-  kwargs: {}
-- name: layer1
-  op: call_module
-  target: layer1
-  args:
-  - norm1
-  kwargs: {}
-- name: flatten1
-  op: call_module
-  target: flatten1
-  args:
-  - layer1
-  kwargs: {}
-- name: output
-  op: output
-  target: output
-  args:
-  - flatten1
-  kwargs: {}
