--- petab_sciml/tmp_examples/data6/nn_model0.yaml
@@ -1,64 +1,0 @@
-nn_model_id: model0
-inputs:
-- input_id: input0
-layers:
-- layer_id: fc1
-  layer_type: Linear
-  args:
-    in_features: 5
-    out_features: 10
-    bias: true
-- layer_id: fc2
-  layer_type: Linear
-  args:
-    in_features: 10
-    out_features: 10
-    bias: true
-- layer_id: fc3
-  layer_type: Linear
-  args:
-    in_features: 10
-    out_features: 2
-    bias: true
-forward:
-- name: input_1
-  op: placeholder
-  target: input
-  args: []
-  kwargs: {}
-- name: fc1
-  op: call_module
-  target: fc1
-  args:
-  - input_1
-  kwargs: {}
-- name: tanh
-  op: call_method
-  target: tanh
-  args:
-  - fc1
-  kwargs: {}
-- name: fc2
-  op: call_module
-  target: fc2
-  args:
-  - tanh
-  kwargs: {}
-- name: tanh_1
-  op: call_method
-  target: tanh
-  args:
-  - fc2
-  kwargs: {}
-- name: fc3
-  op: call_module
-  target: fc3
-  args:
-  - tanh_1
-  kwargs: {}
-- name: output
-  op: output
-  target: output
-  args:
-  - fc3
-  kwargs: {}
